[service] Avoid failures of `service` resource with frequent restarts
Happens on systemd-based systems.

Root of the problem is that the service resource in `datadog_monitor`
is different from the one in the main chef run (chef limitation), so
the restarts that happen there are done immediately instead of being
queued up nicely at the end of the run. Good fix would be to remove
the service definition from `datadog_monitor` and make all invocations
of `datadog_monitor` notify a restart on the global service resource.